Spring add Database create script add javaFX-Spring-Hibernate-PostgresSQL work together now

+++ src/main/java/sample/repositories/FactoryRepository.java
@@ -4,6 +4,6 @@ import sample.entities.FactoryEntity;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface FactoryRepository extends CrudRepository<FactoryEntity, Long>{
+public interface FactoryRepository extends CrudRepository<FactoryEntity, Integer>{
 
 }
